Move `PreloadConfiguration` to the root module
